Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main(){
     const double PI=3.14;
-    int radius;
+    double radius;
     float area;
-    scanf("%d", &radius);
+    scanf("%lf", &radius);
     area = PI * radius * radius;
     printf("Area: %.2f\n", area);
     return 0;
